refactor(editor): move comments plugin into package
